Commit missing routes, models, utils and migrations

+++ alembic/versions/fc20b1fdeb0d_remove_frame_work_field_in_test_reports_.py
@@ -11,6 +11,8 @@ from alembic import op
 import sqlalchemy as sa
 
 
+from sqlalchemy import inspect
+
 # revision identifiers, used by Alembic.
 revision: str = 'fc20b1fdeb0d'
 down_revision: Union[str, Sequence[str], None] = 'a56076ac653b'
@@ -18,8 +20,18 @@ branch_labels: Union[str, Sequence[str], None] = None
 depends_on: Union[str, Sequence[str], None] = None
 
 def upgrade() -> None:
-    op.drop_column('test_reports', 'framework')
+    bind = op.get_bind()
+    inspector = inspect(bind)
+    columns = [c['name'] for c in inspector.get_columns('test_reports')]
+    
+    if 'framework' in columns:
+        op.drop_column('test_reports', 'framework')
 
 
 def downgrade() -> None:
-    op.add_column('test_reports', sa.Column('framework', sa.String(length=255), nullable=True))
+    bind = op.get_bind()
+    inspector = inspect(bind)
+    columns = [c['name'] for c in inspector.get_columns('test_reports')]
+    
+    if 'framework' not in columns:
+        op.add_column('test_reports', sa.Column('framework', sa.String(length=255), nullable=True))
